Removed dead link; Added new PDF (#8796)
* Removed dead link; Added new PDF

This is Visual Basic .Net, and not the old Visual Basic 6 (VB6). Also, this refer to the 2010 version (VB10) latest is the VB16 (2019). There wasn't so many changes since 2010, but some are relevant, like async and await. All changes from 2010 to 2019 can be see at this link: https://learn.microsoft.com/en-us/dotnet/visual-basic/whats-new

* Update free-programming-books-it.md

* Update free-programming-books-it.md